+++ app/scripts/TiledPlot.js
@@ -1995,6 +1995,7 @@ class TiledPlot extends React.Component {
           canvasElement={this.canvasElement}
           centerHeight={this.centerHeight}
           centerWidth={this.centerWidth}
+          disableTrackMenu={this.props.disableTrackMenu}
           dragging={this.props.dragging}
           galleryDim={this.galleryDim}
           height={this.state.height}
@@ -2245,6 +2246,7 @@ TiledPlot.propTypes = {
   canvasElement: PropTypes.object,
   chooseTrackHandler: PropTypes.func,
   chromInfoPath: PropTypes.string,
+  disableTrackMenu: PropTypes.bool,
   dragging: PropTypes.bool,
   editable: PropTypes.bool,
   initialXDomain: PropTypes.array,